added flag to keep old files

+++ dnora/read/ds_read_functions.py
@@ -12,25 +12,26 @@ import geo_parameters as gp
 import ftplib
 
 import os
+
+
 def get_rc_credentials(rc_file: str) -> tuple[str]:
 
-
     # Define the path to your configuration file
-    config_path = os.path.expanduser(f'~/{rc_file}')
+    config_path = os.path.expanduser(f"~/{rc_file}")
 
     # Initialize a dictionary to store the credentials
     config = {}
 
     try:
-        with open(config_path, 'r') as file:
+        with open(config_path, "r") as file:
             for line in file:
                 # Strip whitespace and split the line into key and value
-                key, value = line.strip().split(':', 1)
+                key, value = line.strip().split(":", 1)
                 config[key.strip()] = value.strip()
-        
+
         # Retrieve the username and password
-        ftp_username = config['ftp_username']
-        ftp_password = config['ftp_password']
+        ftp_username = config["ftp_username"]
+        ftp_password = config["ftp_password"]
 
     except FileNotFoundError:
         print(f"Configuration file not found at {config_path}")
@@ -49,21 +50,21 @@ def ftp_read(
     data_type: DnoraDataType,
     data_vars: list[str],
     ## Partial variables in ProductConfiguration
-    **kwargs
-
-):   
+    **kwargs,
+):
     temp_dir = setup_temp_dir(data_type, name)
 
-    username, password = get_rc_credentials('.nchmf_ecmwfrc')
-    ftp = ftplib.FTP('fog.met.no')
+    username, password = get_rc_credentials(".nchmf_ecmwfrc")
+    ftp = ftplib.FTP("fog.met.no")
     ftp.login(username, password)
     ftp.cwd("/")
-    
+
     local_file = f"{temp_dir}/{name}_{url}"
     with open(local_file, "wb") as lf:
         ftp.retrbinary(f"RETR {url}", lf.write)
     ds = xr.open_dataset(local_file)[data_vars]
     return ds.sel(time=slice(start_time, end_time))
+
 
 def find_time_var_in_ds(ds):
     """Tries to identify the time variable in a dataset"""
@@ -222,7 +223,11 @@ def read_one_ds(
                 try_next_file = False
                 keep_trying = False
 
-        except (OSError, RuntimeError, ftplib.error_perm) as e:  # xr gives OSError, fimex gives RuntimeError
+        except (
+            OSError,
+            RuntimeError,
+            ftplib.error_perm,
+        ) as e:  # xr gives OSError, fimex gives RuntimeError
             msg.plain(f'SKIPPING! Got error "{e}" while reading {url}')
             try_next_file = True
 
@@ -372,15 +377,17 @@ def file_is_consistent(
         return False
 
 
-def setup_temp_dir(data_type: DnoraDataType, reader_name: str) -> str:
+def setup_temp_dir(
+    data_type: DnoraDataType, reader_name: str, clean_old_files: bool = True
+) -> str:
     """Sets up a temporery directory for fimex files and cleans out possible old files"""
     temp_folder = f"dnora_{data_type.name.lower()}_temp"
     if not os.path.isdir(temp_folder):
         os.mkdir(temp_folder)
         print("Creating folder %s..." % temp_folder)
-
-    msg.plain("Removing old files from temporary folder...")
-    for f in glob.glob(f"dnora_{data_type.name.lower()}_temp/{reader_name}*.*"):
-        os.remove(f)
+    if clean_old_files:
+        msg.plain("Removing old files from temporary folder...")
+        for f in glob.glob(f"dnora_{data_type.name.lower()}_temp/{reader_name}*.*"):
+            os.remove(f)
 
     return temp_folder
